refactor checl_cla
