+++ .github/workflows/check_cla.yml
@@ -29,12 +29,15 @@ jobs:
         uses: actions/checkout@v3
         with:
           repository: 'dfinity/public-workflows'
+
       - name: Install Python
         uses: actions/setup-python@v4
         with:
           python-version: '3.10'
+
       - name: Install Dependencies
         run: pip install -r requirements.txt
+        
       - name: Check Membership
         id:  check-membership
         run: python reusable_workflows/check_membership/check_membership.py
@@ -54,12 +57,15 @@ jobs:
         uses: actions/checkout@v3
         with:
           repository: 'dfinity/public-workflows'
+
       - name: Install Python
         uses: actions/setup-python@v4
         with:
           python-version: '3.10'
+
       - name: Install Dependencies
         run: pip install -r requirements.txt
+
       - name: Check if accepting external contributions
         id: accepts_external_contrib
         run: |
@@ -96,11 +102,14 @@ jobs:
         uses: actions/checkout@v3
         if: ${{ steps.accepts_external_contrib.outputs.accepts_contrib != 'false' }}
         with:
-          repository: 'dfinity/repositories-open-to-contributions'
+          repository: 'dfinity/public-workflows'
 
       - name: Check CLA
         id:  check-cla
-        uses: ./reusable_workflows/check_cla/check_cla_pr/
+        run: |
+          export PYTHONPATH="$PWD/reusable_workflows/"
+          python reusable_workflows/check_cla/check_cla_pr/check_cla_pr.py
+        shell: bash
         if: ${{ steps.accepts_external_contrib.outputs.accepts_contrib != 'false' }}
         env:
           GH_ORG: ${{ github.repository_owner }}
