--- reusable_workflows/check_cla/check_cla_issue/action.yml
@@ -1,19 +1,0 @@
-name: 'Check CLA'
-description: 'From within the CLA repo, check if the specific CLA issue has been correctly signed.'
-
-runs:
-  using: 'composite'
-  steps:
-    - name: Install Python
-      uses: actions/setup-python@v4
-      with:
-        python-version: '3.10'
-    - name: Install Dependencies
-      run: pip install -q -r requirements.txt
-      shell: bash
-    - name: Check CLA
-      id: check_cla
-      run: |
-        export PYTHONPATH="$PWD/reusable_workflows/"
-        python reusable_workflows/check_cla/check_cla_issue/check_cla_issue.py
-      shell: bash
