--- reusable_workflows/check_cla/check_cla_pr/action.yml
@@ -1,19 +1,0 @@
-name: 'Check CLA'
-description: 'Check if the user is a member or if external, needs to sign CLA agreement. Runs on every PR in the organization.'
-
-runs:
-  using: 'composite'
-  steps:
-    - name: Install Python
-      uses: actions/setup-python@v4
-      with:
-        python-version: '3.10'
-    - name: Install Dependencies
-      run: pip install -q -r requirements.txt
-      shell: bash
-    - name: Check CLA
-      id: check_cla
-      run: |
-        export PYTHONPATH="$PWD/reusable_workflows/"
-        python reusable_workflows/check_cla/check_cla_pr/check_cla_pr.py
-      shell: bash
